Update package.json
Node 8.12.0 is not supported by Alpine 3.8, only by Alpine edge, which is a major problem when installing the latest topo version. The upgrade to enforce Node 8.12 in a minor upgrade breaks builds. Please revert this change (and if applicable, revert the reason for making this change in the first place)